bw-core/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.3.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.11
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.3
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via sqlmodel
sqlmodel==0.0.24
    # via -r requirements.in
starlette==0.46.1
    # via fastapi
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
uvicorn[standard]==0.34.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
